requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv
transformers
torch
Pillow
timm
pymysql